No way to override Primitive Serialization and deserialization of Payload. #108
Added a serializer/deserializer extensibility point, to convert Json
Payload value to Primitive type. And allowed skipping the validation.
